Task 0: Display complex number

+++ 0x00-math_complex/holberton.h
@@ -7,8 +7,8 @@
  */
 typedef struct Complex
 {
-	float re;
-	float im;
+	double re;
+	double im;
 } complex;
 void display_complex_number(complex c);
 #endif
